implement relative path caching
Cache found directories using relative paths only. This allows two mount
namespaces to use the same cache even if their codepath root trees are
different (but point to the same underlying inode).

This also requires that we cache the target dir using its device and
inode numbers.

My use case for this is that on my host, my code is stored in `~/Code`,
but in my containers, I have `/code` symlinked to `~/Code` and find it
nicer to use/display the `/code` path.